Refactor server code to improve modularity and maintainability
- Split functionality into separate provider classes for validation, document symbols, hover, completion, and definitions.
- Introduce CallContextAnalyzer for extracting function call context.
- Add GroovydocParser for extracting and cleaning Groovydoc comments.
- Implement ParameterParser for handling method parameter parsing and matching.
- Remove redundant code and improve organization of server.ts.
- Enhance configuration handling through a dedicated ServerContext class.
