+++ server/src/server.ts
@@ -1,40 +1,28 @@
 import {
   createConnection,
   TextDocuments,
-  Diagnostic,
-  DiagnosticSeverity,
   ProposedFeatures,
   InitializeParams,
   DidChangeConfigurationNotification,
-  CompletionItem,
-  CompletionItemKind,
   TextDocumentPositionParams,
   TextDocumentSyncKind,
   InitializeResult,
   DocumentSymbolParams,
   SymbolInformation,
-  SymbolKind,
-  Location,
-  Range,
-  Position,
   HoverParams,
   Hover,
   DefinitionParams,
   Definition,
-  WorkspaceFolder
+  CompletionItem
 } from 'vscode-languageserver/node';
 
 import { TextDocument } from 'vscode-languageserver-textdocument';
-import * as fs from 'fs';
-import * as path from 'path';
-import {
-  countBracesAndParens,
-  matchesParameterCount,
-  parseMethodParameters,
-  findMatchingParenthesis,
-  findMethodSignature,
-  getSymbolPosition
-} from './utils';
+import { ServerContext } from './core/ServerContext';
+import { ValidationProvider } from './providers/ValidationProvider';
+import { DocumentSymbolProvider } from './providers/DocumentSymbolProvider';
+import { HoverProvider } from './providers/HoverProvider';
+import { CompletionProvider } from './providers/CompletionProvider';
+import { DefinitionProvider } from './providers/DefinitionProvider';
 
 // Create a connection for the server
 const connection = createConnection(ProposedFeatures.all);
@@ -42,28 +30,19 @@ const connection = createConnection(ProposedFeatures.all);
 // Create a simple text document manager
 const documents: TextDocuments<TextDocument> = new TextDocuments(TextDocument);
 
-let hasConfigurationCapability = false;
-let hasWorkspaceFolderCapability = false;
-let workspaceFolders: WorkspaceFolder[] = [];
+// Create server context to manage shared state
+const serverContext = new ServerContext(connection, documents);
 
-// Configuration for Jenkins shared library paths
-let varsPath = 'vars';
-let srcPath = 'src';
+// Create LSP providers
+const validationProvider = new ValidationProvider(serverContext);
+const documentSymbolProvider = new DocumentSymbolProvider(serverContext);
+const hoverProvider = new HoverProvider(serverContext);
+const completionProvider = new CompletionProvider();
+const definitionProvider = new DefinitionProvider(serverContext);
 
 connection.onInitialize((params: InitializeParams) => {
-  const capabilities = params.capabilities;
-
-  hasConfigurationCapability = !!(
-    capabilities.workspace && !!capabilities.workspace.configuration
-  );
-  hasWorkspaceFolderCapability = !!(
-    capabilities.workspace && !!capabilities.workspace.workspaceFolders
-  );
-
-  // Store workspace folders for Jenkins shared library support
-  if (params.workspaceFolders) {
-    workspaceFolders = params.workspaceFolders;
-  }
+  // Initialize server context with capabilities and workspace folders
+  serverContext.initialize(params);
 
   const result: InitializeResult = {
     capabilities: {
@@ -78,7 +57,7 @@ connection.onInitialize((params: InitializeParams) => {
     }
   };
 
-  if (hasWorkspaceFolderCapability) {
+  if (serverContext.hasWorkspaceFolderCapability) {
     result.capabilities.workspace = {
       workspaceFolders: {
         supported: true
@@ -90,848 +69,57 @@ connection.onInitialize((params: InitializeParams) => {
 });
 
 connection.onInitialized(() => {
-  if (hasConfigurationCapability) {
+  if (serverContext.hasConfigurationCapability) {
     connection.client.register(DidChangeConfigurationNotification.type, undefined);
   }
 
   // Request configuration for Jenkins shared library paths
-  if (hasConfigurationCapability) {
+  if (serverContext.hasConfigurationCapability) {
     connection.workspace.getConfiguration('groovy.jenkins.sharedLibrary').then(config => {
-      if (config) {
-        varsPath = config.varsPath || 'vars';
-        srcPath = config.srcPath || 'src';
-        connection.console.log(`Jenkins shared library paths configured: vars=${varsPath}, src=${srcPath}`);
-      }
+      serverContext.updateConfiguration(config);
     });
   }
 });
 
 connection.onDidChangeConfiguration((change) => {
   connection.console.log(`Configuration changed, reloading settings...${JSON.stringify(change, null, 2)}`);
-  if (hasConfigurationCapability) {
+  if (serverContext.hasConfigurationCapability) {
     connection.workspace.getConfiguration('groovy.jenkins.sharedLibrary').then(config => {
-      if (config) {
-        varsPath = config.varsPath || 'vars';
-        srcPath = config.srcPath || 'src';
-        connection.console.log(`Jenkins shared library paths updated: vars=${varsPath}, src=${srcPath}`);
-      }
+      serverContext.updateConfiguration(config);
     });
   }
 });
 
 // The content of a text document has changed
 documents.onDidChangeContent(change => {
-  validateTextDocument(change.document);
+  validationProvider.validateTextDocument(change.document);
 });
-
-async function validateTextDocument(textDocument: TextDocument): Promise<void> {
-  const text = textDocument.getText();
-  const diagnostics: Diagnostic[] = [];
-
-  // Basic syntax validation
-  const counts = countBracesAndParens(text);
-  
-  if (counts.openBraces !== counts.closeBraces) {
-    const diagnostic: Diagnostic = {
-      severity: DiagnosticSeverity.Error,
-      range: {
-        start: textDocument.positionAt(0),
-        end: textDocument.positionAt(text.length)
-      },
-      message: `Unmatched braces: ${counts.openBraces} opening, ${counts.closeBraces} closing`,
-      source: 'groovy'
-    };
-    diagnostics.push(diagnostic);
-  }
-
-  if (counts.openParens !== counts.closeParens) {
-    const diagnostic: Diagnostic = {
-      severity: DiagnosticSeverity.Error,
-      range: {
-        start: textDocument.positionAt(0),
-        end: textDocument.positionAt(text.length)
-      },
-      message: `Unmatched parentheses: ${counts.openParens} opening, ${counts.closeParens} closing`,
-      source: 'groovy'
-    };
-    diagnostics.push(diagnostic);
-  }
-
-  connection.sendDiagnostics({ uri: textDocument.uri, diagnostics });
-}
 
 // Document symbols (outline)
 connection.onDocumentSymbol((params: DocumentSymbolParams): SymbolInformation[] => {
-  const document = documents.get(params.textDocument.uri);
-  if (!document) {
-    return [];
-  }
-
-  const text = document.getText();
-  const symbols: SymbolInformation[] = [];
-
-  // Match class definitions
-  const classRegex = /(?:^|\n)\s*(?:public|private|protected)?\s*(?:abstract|final)?\s*class\s+(\w+)/g;
-  let match;
-  while ((match = classRegex.exec(text)) !== null) {
-    const className = match[1];
-    const position = document.positionAt(match.index);
-    symbols.push({
-      name: className,
-      kind: SymbolKind.Class,
-      location: Location.create(
-        document.uri,
-        Range.create(position, Position.create(position.line, position.character + className.length))
-      )
-    });
-  }
-
-  // Match method definitions
-  const methodRegex = /(?:^|\n)\s*(?:public|private|protected)?\s*(?:static)?\s*(?:def|void|\w+)\s+(\w+)\s*\(/g;
-  while ((match = methodRegex.exec(text)) !== null) {
-    const methodName = match[1];
-    const position = document.positionAt(match.index);
-    symbols.push({
-      name: methodName,
-      kind: SymbolKind.Method,
-      location: Location.create(
-        document.uri,
-        Range.create(position, Position.create(position.line, position.character + methodName.length))
-      )
-    });
-  }
-
-  // Match property/field definitions
-  const propertyRegex = /(?:^|\n)\s*(?:public|private|protected)?\s*(?:static|final)?\s*(?:def|\w+)\s+(\w+)\s*=/g;
-  while ((match = propertyRegex.exec(text)) !== null) {
-    const propName = match[1];
-    const position = document.positionAt(match.index);
-    symbols.push({
-      name: propName,
-      kind: SymbolKind.Property,
-      location: Location.create(
-        document.uri,
-        Range.create(position, Position.create(position.line, position.character + propName.length))
-      )
-    });
-  }
-
-  return symbols;
+  return documentSymbolProvider.getDocumentSymbols(params);
 });
 
 // Hover information
 connection.onHover((params: HoverParams): Hover | null => {
-  const document = documents.get(params.textDocument.uri);
-  if (!document) {
-    return null;
-  }
-
-  const text = document.getText();
-  const offset = document.offsetAt(params.position);
-
-  // First, try to extract function call context for signature information
-  const functionCallContext = extractFunctionCallContext(text, offset);
-  if (functionCallContext) {
-    const { symbol, fullSymbol, args } = functionCallContext;
-
-    // Search for matching function signature
-    const signatureCurrentDoc = findFunctionSignatureInDocument(text, symbol, args);
-    if (signatureCurrentDoc) {
-      return {
-        contents: {
-          kind: 'markdown',
-          value: signatureCurrentDoc
-        }
-      };
-    }
-
-    // If no exact signature match, try to find any signature info for the symbol in workspace
-    const signatureWorkspace = findFunctionSignatureInWorkspace(symbol, fullSymbol, args);
-    if (signatureWorkspace) {
-      return {
-        contents: {
-          kind: 'markdown',
-          value: signatureWorkspace
-        }
-      };
-    }
-  }
-
-  // Find the word at the cursor position for keyword/symbol info
-  const wordPattern = /\b\w+\b/g;
-  let match;
-  while ((match = wordPattern.exec(text)) !== null) {
-    if (match.index <= offset && offset <= match.index + match[0].length) {
-      const word = match[0];
-
-      // Provide hover info for Groovy keywords
-      const keywordInfo = getKeywordInfo(word);
-      if (keywordInfo) {
-        return {
-          contents: {
-            kind: 'markdown',
-            value: keywordInfo
-          }
-        };
-      }
-
-      // Check if it's a class, method, or property
-      const doc = extractGroovydocForSymbolInText(text, word);
-      if (doc) {
-        return {
-          contents: {
-            kind: 'markdown',
-            value: doc
-          }
-        };
-      }
-
-      const symbolInfo = findSymbolInfo(text, word);
-      if (symbolInfo) {
-        return {
-          contents: {
-            kind: 'markdown',
-            value: symbolInfo
-          }
-        };
-      }
-    }
-  }
-
-  return null;
+  return hoverProvider.getHover(params);
 });
-
-function getKeywordInfo(keyword: string): string | null {
-  const keywords: { [key: string]: string } = {
-    'def': '**def** - Defines a variable or method with dynamic typing',
-    'class': '**class** - Defines a class',
-    'interface': '**interface** - Defines an interface',
-    'extends': '**extends** - Specifies class inheritance',
-    'implements': '**implements** - Specifies interface implementation',
-    'import': '**import** - Imports classes or packages',
-    'package': '**package** - Declares the package',
-    'return': '**return** - Returns a value from a method',
-    'if': '**if** - Conditional statement',
-    'else': '**else** - Alternative branch of conditional',
-    'for': '**for** - Loop statement',
-    'while': '**while** - Loop statement',
-    'switch': '**switch** - Multi-way branch statement',
-    'case': '**case** - Branch in switch statement',
-    'break': '**break** - Exits loop or switch',
-    'continue': '**continue** - Skips to next iteration',
-    'try': '**try** - Begins exception handling block',
-    'catch': '**catch** - Catches exceptions',
-    'finally': '**finally** - Always executed block',
-    'throw': '**throw** - Throws an exception',
-    'new': '**new** - Creates a new instance',
-    'this': '**this** - References current object',
-    'super': '**super** - References parent class',
-    'static': '**static** - Defines class-level member',
-    'final': '**final** - Makes variable or class immutable',
-    'public': '**public** - Public access modifier',
-    'private': '**private** - Private access modifier',
-    'protected': '**protected** - Protected access modifier'
-  };
-
-  return keywords[keyword] || null;
-}
-
-function findFunctionSignatureInDocument(text: string, symbol: string, args: string[]): string | null {
-  // Try to find method definition with matching signature
-  const signature = findMethodSignature(text, symbol, args);
-  if (signature) {
-    const doc = extractGroovydocForSymbolInText(text, symbol, args);
-    return formatSignature(signature, doc);
-  }
-
-  return null;
-}
-
-function findFunctionSignatureInWorkspace(symbol: string, fullSymbol: string, args: string[]): string | null {
-  if (workspaceFolders.length === 0) {
-    return null;
-  }
-
-  for (const workspaceFolder of workspaceFolders) {
-    // Search in src/ directory for classes and their methods
-    const srcDir = path.join(workspaceFolder.uri.replace('file://', ''), srcPath);
-    if (fs.existsSync(srcDir)) {
-      const signature = findMethodSignatureInSrc(srcDir, symbol, args);
-      if (signature) {
-        return signature;
-      }
-    }
-    // Search in vars/ directory for global functions (Jenkins shared library)
-    const varsDir = path.join(workspaceFolder.uri.replace('file://', ''), varsPath);
-    if (fs.existsSync(varsDir)) {
-      try {
-        const files = fs.readdirSync(varsDir);
-        
-        // If fullSymbol contains a dot, it's a qualified call like myUtils.deployTo()
-        // Extract the object/namespace part and prioritize that file
-        const symbolParts = fullSymbol.split('.');
-        const objectName = symbolParts.length > 1 ? symbolParts[0] : null;
-        
-        // First pass: if it's a qualified call, look in the specific file
-        if (objectName) {
-          const targetFile = `${objectName}.groovy`;
-          const targetFilePath = path.join(varsDir, targetFile);
-          if (fs.existsSync(targetFilePath)) {
-            const content = fs.readFileSync(targetFilePath, 'utf8');
-            const signature = findMethodSignature(content, symbol, args);
-            if (signature) {
-              const doc = extractGroovydocForSymbolInText(content, symbol, args);
-              return formatSignature(signature, doc);
-            }
-          }
-        }
-        
-        // Second pass: search all files (for unqualified calls or as fallback)
-        for (const file of files) {
-          if (file.endsWith('.groovy')) {
-            const filePath = path.join(varsDir, file);
-            const content = fs.readFileSync(filePath, 'utf8');
-
-            // In Jenkins shared libraries, vars files can be called by their filename
-            const fileNameWithoutExt = file.replace('.groovy', '');
-            if (fileNameWithoutExt === symbol) {
-              // If the symbol matches the filename, look for the 'call' function with matching signature
-              const signature = findMethodSignature(content, 'call', args);
-              if (signature) {
-                    const doc = extractGroovydocForSymbolInText(content, 'call', args);
-                    return formatSignature(signature, doc);
-              }
-            } else if (!objectName || fileNameWithoutExt !== objectName) {
-              // Only look for the function if we haven't already checked this file in the first pass
-              const signature = findMethodSignature(content, symbol, args);
-              if (signature) {
-                    const doc = extractGroovydocForSymbolInText(content, symbol, args);
-                    return formatSignature(signature, doc);
-              }
-            }
-          }
-        }
-      } catch (error) {
-        connection.console.log(`Error reading vars directory: ${error}`);
-      }
-    }
-  }
-
-  return null;
-}
-
-function findMethodSignatureInSrc(srcDir: string, symbol: string, args: string[]): string | null {
-  // Recursively search for method signature in src directory
-  function searchDirectory(dir: string): string | null {
-    try {
-      const items = fs.readdirSync(dir);
-      for (const item of items) {
-        const itemPath = path.join(dir, item);
-        const stat = fs.statSync(itemPath);
-
-        if (stat.isDirectory()) {
-          const result = searchDirectory(itemPath);
-          if (result) {
-            return result
-          };
-        } else if (item.endsWith('.groovy')) {
-          const content = fs.readFileSync(itemPath, 'utf8');
-          const signature = findMethodSignature(content, symbol, args);
-          const doc = extractGroovydocForSymbolInText(content, symbol, args);
-          if (signature) {
-            return formatSignature(signature, doc);
-          }
-        }
-      }
-    } catch (error) {
-      connection.console.log(`Error reading src directory: ${error}`);
-    }
-    return null;
-  }
-
-  return searchDirectory(srcDir);
-}
-
-function formatSignature(signature: string, doc?: string): string {
-  // Clean up the signature for display
-  const cleanSignature = signature.replace(/\s+/g, ' ').trim();
-
-  let md = '';
-  if (doc) {
-    const cleanedDoc = doc.trim();
-    md += cleanedDoc + '\n\n';
-  }
-  md += `\`\`\`groovy\n${cleanSignature}\n\`\`\``;
-  return md;
-}
-
-function extractGroovydocForSymbolInText(text: string, symbol: string, args?: string[]): string | undefined {
-  // Search for common definition patterns and look for a /** ... */ block immediately above
-  // If args are provided, match the correct overload by parameter count
-  const defPatterns = [
-    new RegExp(`(?:^|\\n)\\s*(?:public|private|protected)?\\s*(?:abstract|final)?\\s*(?:class|interface|enum|trait)\\s+${symbol}\\b`, 'g'),
-    new RegExp(`(?:^|\\n)\\s*(?:public|private|protected)?\\s*(?:static)?\\s*(?:def|void|\\w+(?:<[^>]*>)?)\\s+${symbol}\\s*\\(`, 'g'),
-    new RegExp(`(?:^|\\n)\\s*(?:def\\s+)?${symbol}\\s*\\(`, 'g')
-  ];
-
-  for (const pattern of defPatterns) {
-    let m;
-    while ((m = pattern.exec(text)) !== null) {
-      const defIndex = m.index;
-
-      const commentEnd = text.lastIndexOf('*/', defIndex);
-      if (commentEnd === -1) {continue;}
-      const commentStart = text.lastIndexOf('/**', commentEnd);
-      if (commentStart === -1) {continue;}
-
-      // Ensure only whitespace/annotations between comment end and definition
-      const between = text.substring(commentEnd + 2, defIndex);
-      if (!/^[\s@]*$/.test(between)) {continue;}
-
-      // If args are provided, check if this definition matches the parameter count (overload matching)
-      if (args !== undefined && m[0].indexOf('(') !== -1) {
-        // This looks like a method definition, extract parameter list
-        const paramStart = defIndex + m[0].length - 1; // Position of opening paren
-        const paramEnd = findMatchingParenthesis(text, paramStart);
-        if (paramEnd !== -1) {
-          const paramList = text.substring(paramStart + 1, paramEnd);
-          const expectedParams = parseMethodParameters(paramList);
-          
-          // Check if parameter count matches (considering default parameters)
-          if (!matchesParameterCount(expectedParams, args.length)) {
-            continue; // This overload doesn't match, try next
-          }
-        }
-      }
-
-      const raw = text.substring(commentStart, commentEnd + 2);
-      return cleanGroovydoc(raw);
-    }
-  }
-
-  return undefined;
-}
-
-function cleanGroovydoc(raw: string): string {
-  // Strip /** and */ and leading * on each line
-  let inner = raw.replace(/^\/\*\*/s, '').replace(/\*\/$/s, '');
-  const lines = inner.split(/\r?\n/).map(l => l.replace(/^\s*\*\s?/, '')).map(l => l.replace(/^\s+|\s+$/g, ''));
-  return lines.join('\n').trim();
-}
-
-function findSymbolInfo(text: string, symbol: string): string | null {
-  // Check if it's a class
-  const classRegex = new RegExp(`class\\s+${symbol}\\s*(?:extends|implements|\\{)`, 'g');
-  if (classRegex.test(text)) {
-    return `**Class**: ${symbol}`;
-  }
-
-  // Check if it's a method
-  const methodRegex = new RegExp(`(?:def|void|\\w+)\\s+${symbol}\\s*\\(`, 'g');
-  if (methodRegex.test(text)) {
-    return `**Method**: ${symbol}`;
-  }
-
-  return null;
-}
 
 // Code completion
 connection.onCompletion(
-  (_textDocumentPosition: TextDocumentPositionParams): CompletionItem[] => {
-    const groovyKeywords = [
-      'abstract', 'as', 'assert', 'break', 'case', 'catch', 'class', 'const',
-      'continue', 'def', 'default', 'do', 'else', 'enum', 'extends', 'false',
-      'final', 'finally', 'for', 'goto', 'if', 'implements', 'import', 'in',
-      'instanceof', 'interface', 'new', 'null', 'package', 'private', 'protected',
-      'public', 'return', 'static', 'super', 'switch', 'this', 'throw', 'throws',
-      'trait', 'true', 'try', 'while'
-    ];
-
-    return groovyKeywords.map(keyword => ({
-      label: keyword,
-      kind: CompletionItemKind.Keyword,
-      data: keyword
-    }));
+  (params: TextDocumentPositionParams): CompletionItem[] => {
+    return completionProvider.getCompletions(params);
   }
 );
 
 connection.onCompletionResolve((item: CompletionItem): CompletionItem => {
-  const keywordDescriptions: { [key: string]: string } = {
-    'def': 'Defines a variable or method with dynamic typing',
-    'class': 'Defines a class',
-    'interface': 'Defines an interface',
-    'trait': 'Defines a trait (mixin)',
-    'return': 'Returns a value from a method',
-    'if': 'Conditional statement',
-    'for': 'Loop statement',
-    'while': 'Loop statement',
-    'switch': 'Multi-way branch statement',
-    'try': 'Exception handling block'
-  };
-
-  if (item.data && keywordDescriptions[item.data]) {
-    item.detail = keywordDescriptions[item.data];
-    item.documentation = keywordDescriptions[item.data];
-  }
-
-  return item;
+  return completionProvider.resolveCompletionItem(item);
 });
 
 // Go to Definition
 connection.onDefinition((params: DefinitionParams): Definition | null => {
-  const document = documents.get(params.textDocument.uri);
-  if (!document) {
-    return null;
-  }
-
-  const text = document.getText();
-  const offset = document.offsetAt(params.position);
-
-  // Find the word at the cursor position and extract function call context
-  const functionCallContext = extractFunctionCallContext(text, offset);
-  if (!functionCallContext) {
-    return null;
-  }
-
-  const { symbol, fullSymbol, args } = functionCallContext;
-
-  // First, search for the definition in the current document
-  let definition = findDefinitionInDocumentWithSignature(document, text, symbol, args);
-  if (definition) {
-    return definition;
-  }
-
-  // If not found in current document, search in Jenkins shared library files
-  definition = findDefinitionInJenkinsSharedLibraryWithSignature(symbol, fullSymbol, args);
-  return definition;
+  return definitionProvider.getDefinition(params);
 });
-
-function extractFunctionCallContext(text: string, offset: number): { symbol: string, fullSymbol: string, args: string[] } | null {
-  // Find the function call that contains the offset
-  // Look for patterns like: symbol(args) or obj.symbol(args)
-
-  // First, try to find a function call ending at or after the offset
-  const functionCallPattern = /\b(?!if\b|while\b|for\b|switch\b|catch\b|return\b)(\w+(?:\.\w+)*)\s*\(((?:[^()]+|\((?:[^()]+|\([^()]*\))*\))*)\)/g;
-  let match;
-
-  while ((match = functionCallPattern.exec(text)) !== null) {
-    const callStart = match.index;
-    const callEnd = match.index + match[0].length;
-
-    // Check if the offset is within this function call
-    if (callStart <= offset && offset <= callEnd) {
-      const fullSymbol = match[1];
-      const argsString = match[2];
-
-      // Extract the actual function name (last part after dots)
-      const symbolParts = fullSymbol.split('.');
-      const symbol = symbolParts[symbolParts.length - 1];
-
-      // Parse arguments
-      const args = parseArguments(argsString);
-
-      return { symbol, fullSymbol, args };
-    }
-  }
-
-  return null;
-}
-
-function parseArguments(argsString: string): string[] {
-  if (!argsString.trim()) {
-    return [];
-  }
-
-  // Simple argument parsing - split by commas but be careful with nested structures
-  const args: string[] = [];
-  let currentArg = '';
-  let parenDepth = 0;
-  let bracketDepth = 0;
-  let inString = false;
-  let stringChar = '';
-
-  for (let i = 0; i < argsString.length; i++) {
-    const char = argsString[i];
-
-    if (inString) {
-      currentArg += char;
-      if (char === stringChar && argsString[i - 1] !== '\\') {
-        inString = false;
-        stringChar = '';
-      }
-    } else {
-      if ((char === '"' || char === "'") && argsString[i - 1] !== '\\') {
-        inString = true;
-        stringChar = char;
-        currentArg += char;
-      } else if (char === '(') {
-        parenDepth++;
-        currentArg += char;
-      } else if (char === ')') {
-        parenDepth--;
-        currentArg += char;
-      } else if (char === '[') {
-        bracketDepth++;
-        currentArg += char;
-      } else if (char === ']') {
-        bracketDepth--;
-        currentArg += char;
-      } else if (char === ',' && parenDepth === 0 && bracketDepth === 0) {
-        args.push(currentArg.trim());
-        currentArg = '';
-      } else {
-        currentArg += char;
-      }
-    }
-  }
-
-  if (currentArg.trim()) {
-    args.push(currentArg.trim());
-  }
-
-  return args;
-}
-
-function findDefinitionInDocumentWithSignature(document: TextDocument, text: string, symbol: string, args: string[]): Location | null {
-  // Try to find class definition
-  const classRegex = new RegExp(`(?:^|\\n)\\s*(?:public|private|protected)?\\s*(?:abstract|final)?\\s*(?:class|interface|enum|trait)\\s+(${symbol})\\b`, 'g');
-  let match = classRegex.exec(text);
-  if (match) {
-    const { line, character } = getSymbolPosition(match, symbol, text);
-    return Location.create(document.uri, Range.create(
-      Position.create(line, character),
-      Position.create(line, character + symbol.length)
-    ));
-  }
-
-  // Try to find method definition with matching signature
-  const location = findMethodDefinitionWithSignature(text, symbol, args);
-  if (location) {
-    return Location.create(document.uri, location.range);
-  }
-
-  // Try to find property/field definition
-  const propertyRegex = new RegExp(`(?:^|\\n)\\s*(?:public|private|protected)?\\s*(?:static|final)?\\s*(?:def|\\w+)\\s+(${symbol})\\s*=`, 'g');
-  match = propertyRegex.exec(text);
-  if (match) {
-    const { line, character } = getSymbolPosition(match, symbol, text);
-    return Location.create(document.uri, Range.create(
-      Position.create(line, character),
-      Position.create(line, character + symbol.length)
-    ));
-  }
-
-  // Try to find variable definition (including method parameters and local variables)
-  const variableRegex = new RegExp(`(?:^|\\n|\\(|,)\\s*(?:def|\\w+)?\\s+(${symbol})\\s*(?:=|,|\\)|\\n)`, 'g');
-  match = variableRegex.exec(text);
-  if (match) {
-    const { line, character } = getSymbolPosition(match, symbol, text);
-    return Location.create(document.uri, Range.create(
-      Position.create(line, character),
-      Position.create(line, character + symbol.length)
-    ));
-  }
-
-  return null;
-}
-
-function findMethodDefinitionWithSignature(text: string, symbol: string, args: string[]): Location | null {
-  const methodRegex = new RegExp(`(?:^|\\n)\\s*(?:public|private|protected)?\\s*(?:static)?\\s*(?:def|void|\\w+(?:<[^>]*>)?(?:\\s*<[^>]*>)*)\\s+${symbol}\\s*\\(`, 'g');
-
-  let match;
-  while ((match = methodRegex.exec(text)) !== null) {
-    // Extract the parameter list from the method definition
-    const paramStart = match.index + match[0].length - 1; // Position after opening parenthesis
-    const paramEnd = findMatchingParenthesis(text, paramStart);
-    if (paramEnd !== -1) {
-      const paramList = text.substring(paramStart + 1, paramEnd);
-      const expectedParams = parseMethodParameters(paramList);
-
-      // Check if parameter count matches (considering default parameters)
-      if (matchesParameterCount(expectedParams, args.length)) {
-        // Found a matching signature
-        // Find the actual position of the symbol within the match string
-        const { line, character } = getSymbolPosition(match, symbol, text);
-        
-        return Location.create(
-          '', // Will be set by caller
-          Range.create(
-            Position.create(line, character),
-            Position.create(line, character + symbol.length)
-          )
-        );
-      }
-    }
-  }
-
-  return null;
-}
-
-function findDefinitionInJenkinsSharedLibraryWithSignature(symbol: string, fullSymbol: string, args: string[]): Location | null {
-  if (workspaceFolders.length === 0) {
-    return null;
-  }
-
-  // Search in vars/ directory for global functions
-  for (const workspaceFolder of workspaceFolders) {
-    const varsDir = path.join(workspaceFolder.uri.replace('file://', ''), varsPath);
-    if (fs.existsSync(varsDir)) {
-      connection.console.log(`Searching for function ${symbol} with ${args.length} args in vars directory: ${varsDir}`);
-      try {
-        const files = fs.readdirSync(varsDir);
-        
-        // If fullSymbol contains a dot, it's a qualified call like myUtils.deployTo()
-        // Extract the object/namespace part and prioritize that file
-        const symbolParts = fullSymbol.split('.');
-        const objectName = symbolParts.length > 1 ? symbolParts[0] : null;
-        
-        // First pass: if it's a qualified call, look in the specific file
-        if (objectName) {
-          const targetFile = `${objectName}.groovy`;
-          const targetFilePath = path.join(varsDir, targetFile);
-          if (fs.existsSync(targetFilePath)) {
-            const content = fs.readFileSync(targetFilePath, 'utf8');
-            const location = findFunctionDefinitionWithSignature(targetFilePath, content, symbol, args);
-            if (location) {
-              return location;
-            }
-          }
-        }
-        
-        // Second pass: search all files (for unqualified calls or as fallback)
-        for (const file of files) {
-          if (file.endsWith('.groovy')) {
-            const filePath = path.join(varsDir, file);
-            const content = fs.readFileSync(filePath, 'utf8');
-
-            // In Jenkins shared libraries, vars files can be called by their filename
-            // e.g., myUtils() calls the 'call' function in myUtils.groovy
-            const fileNameWithoutExt = file.replace('.groovy', '');
-            if (fileNameWithoutExt === symbol) {
-              // If the symbol matches the filename, look for the 'call' function with matching signature
-              const location = findFunctionDefinitionWithSignature(filePath, content, 'call', args);
-              if (location) {
-                return location;
-              }
-            } else if (!objectName || fileNameWithoutExt !== objectName) {
-              // Only look for the function if we haven't already checked this file in the first pass
-              const location = findFunctionDefinitionWithSignature(filePath, content, symbol, args);
-              if (location) {
-                return location;
-              }
-            }
-          }
-        }
-      } catch (error) {
-        connection.console.log(`Error reading vars directory: ${error}`);
-      }
-    }
-
-    // Search in src/ directory for classes and their methods
-    const srcDir = path.join(workspaceFolder.uri.replace('file://', ''), srcPath);
-    if (fs.existsSync(srcDir)) {
-      connection.console.log(`Searching for class or method ${symbol} with ${args.length} args in src directory: ${srcDir}`);
-      const location = findClassOrMethodDefinitionInSrcWithSignature(srcDir, symbol, args);
-      if (location) {
-        return location;
-      }
-    }
-  }
-
-  return null;
-}
-
-function findFunctionDefinitionWithSignature(filePath: string, text: string, symbol: string, args: string[]): Location | null {
-  const functionPatterns = [
-    // Standard function definition with def
-    new RegExp(`(?:^|\\n)\\s*(?:def\\s+)?${symbol}\\s*\\(`, 'g'),
-    // Function without def keyword (direct function definition)
-    new RegExp(`(?:^|\\n)\\s*${symbol}\\s*\\(`, 'g')
-  ];
-
-  for (const functionRegex of functionPatterns) {
-    let match;
-    while ((match = functionRegex.exec(text)) !== null) {
-      // Extract the parameter list from the function definition
-      const paramStart = match.index + match[0].length - 1; // Position after opening parenthesis
-      const paramEnd = findMatchingParenthesis(text, paramStart);
-      if (paramEnd !== -1) {
-        const paramList = text.substring(paramStart + 1, paramEnd);
-        const expectedParams = parseMethodParameters(paramList);
-
-        // Check if parameter count matches (considering default parameters)
-        if (matchesParameterCount(expectedParams, args.length)) {
-          // Found a matching signature
-          // Find the actual position of the symbol within the match string
-          const { line, character } = getSymbolPosition(match, symbol, text);
-          
-          return Location.create(
-            `file://${filePath}`,
-            Range.create(
-              Position.create(line, character),
-              Position.create(line, character + symbol.length)
-            )
-          );
-        }
-      }
-    }
-  }
-
-  return null;
-}
-
-function findClassOrMethodDefinitionInSrcWithSignature(srcDir: string, symbol: string, args: string[]): Location | null {
-  // Recursively search for class or method definition in src directory
-  function searchDirectory(dir: string): Location | null {
-    try {
-      const items = fs.readdirSync(dir);
-      for (const item of items) {
-        const itemPath = path.join(dir, item);
-        const stat = fs.statSync(itemPath);
-
-        if (stat.isDirectory()) {
-          const result = searchDirectory(itemPath);
-          if (result) {
-            return result
-          };
-        } else if (item.endsWith('.groovy')) {
-          const content = fs.readFileSync(itemPath, 'utf8');
-          // First try to find class definition
-          let location = findClassDefinitionInFile(itemPath, content, symbol);
-          if (location) {
-            return location;
-          }
-          // Then try to find method definition with matching signature
-          location = findMethodDefinitionWithSignature(content, symbol, args);
-          if (location) {
-            return Location.create(`file://${itemPath}`, location.range);
-          }
-        }
-      }
-    } catch (error) {
-      connection.console.log(`Error reading directory: ${error}`);
-    }
-    return null;
-  }
-
-  return searchDirectory(srcDir);
-}
-
-function findClassDefinitionInFile(filePath: string, content: string, symbol: string): Location | null {
-  const classRegex = new RegExp(`(?:^|\\n)\\s*(?:public|private|protected)?\\s*(?:abstract|final)?\\s*(?:class|interface|enum|trait)\\s+${symbol}\\b`, 'g');
-  const match = classRegex.exec(content);
-  if (match) {
-    const { line, character } = getSymbolPosition(match, symbol, content);
-    return Location.create(
-      `file://${filePath}`,
-      Range.create(
-        Position.create(line, character),
-        Position.create(line, character + symbol.length)
-      )
-    );
-  }
-  return null;
-}
 
 documents.listen(connection);
